Use Keycloak as Auth tool

+++ src/main/resources/db/migration/V1.0.1__Quarkus.sql
@@ -1,16 +1,16 @@
 CREATE TABLE person (
-    id BIGINT PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE project (
-    id BIGINT PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE project_member (
-    project_id BIGINT NOT NULL,
-    person_id BIGINT NOT NULL,
+    project_id INT NOT NULL,
+    person_id INT NOT NULL,
     PRIMARY KEY (project_id, person_id),
     FOREIGN KEY (project_id) REFERENCES project (id),
     FOREIGN KEY (person_id) REFERENCES person (id)
